For use with Title Seacher Program
The titles searcher program, extract3.py, uses all of these text files. Additionally, the titles searcher program makes use of a title scraper that is available to the public on Github, https://github.com/NikolaiT/GoogleScraper. To use the titles searcher program, you will need to download "GoogleScraper" from the aforementioned link. You will need to create a file called keywords.txt with a list of all of the websites that you would like to get the titles for. Then put the following command into a command terminal to generate output for the titles searcher program to analyze: "GoogleScraper -s "bing" -m http --keyword-file keywords.txt  >output.txt". In this command keywords.txt is the name of the text file with the list of all of the website (or whatever name you give it) and output.txt is the name of the output file for the scraper results. You must open extract3.py and change the name of the file referenced in "pin1 = open('out3.txt','r')" to whatever the name of the ou